backend/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
# For future PostgreSQL upgrade (optional, SQLite used for MVP):
# psycopg2-binary==2.9.9
# For Redis/Celery (skipped for MVP to save cost):
# redis==5.0.1
# celery==5.3.4
